+++ Video Capture SDK/WinForms/CSharp/madVR demo/Form1.cs
@@ -150,7 +150,7 @@ namespace madVR_Demo
                     return;
                 }
 
-                var videoFormat = deviceItem.VideoFormats.FirstOrDefault(format => format.Name == cbVideoInputFormat.Text);
+                var videoFormat = deviceItem.VideoFormats.Find(format => format.Name == cbVideoInputFormat.Text);
                 if (videoFormat == null)
                 {
                     return;